% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petmart.R
\name{petmart}
\alias{petmart}
\title{Take some pageviews, output Markov model prediction}
\usage{
petmart(current_url)
}
\arguments{
\item{current_url}{the url to predict from}
}
\value{
The prediction
}
\description{
Take some pageviews, output Markov model prediction
}
