% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictClicks.R
\name{predictClicks}
\alias{predictClicks}
\title{Take some pageviews, output Markov model prediction}
\usage{
predictClicks(current_url)
}
\arguments{
\item{current_url}{the url to predict from}
}
\value{
The prediction
}
\description{
Take some pageviews, output Markov model prediction
}
