% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMarkov.R
\name{predictPetmart}
\alias{predictPetmart}
\title{Predict next page model 2}
\usage{
predictPetmart(current_url)
}
\arguments{
\item{current_url}{the url to predict from}
}
\description{
Predict next page model 2
}
